third_party/WebKit/LayoutTests/svg/dom/SVGLengthList-setEmptyString.html

<!DOCTYPE HTML>
<title>SVGLengthList, set empty string</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<svg width="200" height="200">
    <text></text>
</svg>
<script>
test(function() {
  var text = document.querySelector('text');
  var list = text.x.baseVal;
  text.setAttribute("x", "1 2 3");
  assert_equals(list.numberOfItems, 3);
  text.setAttribute("x", "");
  assert_equals(list.numberOfItems, 0);

  text.setAttribute("x", "1 2 3");
  assert_equals(list.numberOfItems, 3);
  text.removeAttribute("x");
  assert_equals(list.numberOfItems, 0);
});
</script>
